Multiple fixes:
added a setup.py to allow armor_unet to be used a package and import to anywhere within the environment of the project.

altered the training loop to allow for argparse instead on evnironment variable editing (that was stupid ngl). added the ability to name each wandb run with a flag "--run-name __" as well as flags for epochs, batch size, and learning rate per run.

side addition: I finally added the optimization for RTX cards for targeting the matmul percison of the tensor cores. it was literally just:
torch.set_float32_matmul_precision() with 'high' or 'medium'